% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivBootnet.R
\name{IndivBootnet}
\alias{IndivBootnet}
\title{Perform permutation test on the individual static or temporal network using bootstrap.}
\usage{
IndivBootnet(
  data,
  IsingResult,
  id,
  type = c("nonparametric", "casedropping"),
  sample_prob = seq(0.9, 0.5, -0.1),
  nBoots = 500,
  nCores = 1,
  verbose = TRUE,
  library = .libPaths(),
  seed
)
}
\arguments{
\item{data}{The data frame used to fit the IndivIsing object.}

\item{IsingResult}{The fitted IndivIsing object.}

\item{id}{The column name in 'data' that denotes the ID of each subject.}

\item{type}{The type of bootstrap procedure. Can be either "nonparametric" for edge testing or
"casedropping" for centrality testing.}

\item{sample_prob}{The vector of sampling proportion for casedropping bootstrap. Default to (0.9,0.8,...0.5).}

\item{nBoots}{The number of bootstraps. Defaults to 500.}

\item{nCores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.
Defaults to 1.}

\item{verbose}{Logical. Should progress of the function be printed to the console?
Defaults to TRUE.}

\item{library}{Library location to be used in parallel computing. Default to .libPaths().}

\item{seed}{The random seed.}
}
\value{
An IndivBootnet object.

\code{bootResults} A list containing the bootstrap results. A list of fitted networks for
"nonparametric" bootstrap, and a list of centrality correlation values for "casedropping" bootstrap.

\code{id} The input id variable.

\code{type} The input type of bootstrap.

\code{sample_prob} The input sampling proportion for casedropping bootstrap.

\code{time} Time spent on running the function.
}
\description{
This function performs permutation tests on the constructed individual static or
temporal network using bootstrap to evaluate its accuracy and stability. A nonparametric
bootstrap sampled with replacement is performed to assess the accuracy of estimated
network connections. A case-dropping bootstrap sampled without replacement is performed to
assess the stability of the calculated centrality indices.
}
