% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivBootSplitting.R
\name{IndivBootSplitting}
\alias{IndivBootSplitting}
\title{Perform data splitting inference by permutation test on the individual static or temporal network.}
\usage{
IndivBootSplitting(
  data,
  IsingResult,
  id,
  type = c("nonparametric", "casedropping"),
  sample_prob = seq(0.9, 0.5, -0.1),
  nBoots = 500,
  nCores = 1,
  verbose = TRUE,
  library = .libPaths(),
  seed
)
}
\arguments{
\item{data}{The independent data frame used for data splitting inference.}

\item{IsingResult}{The fitted IndivIsing object.}

\item{id}{The column name in 'data' that indicates the ID of each subject.}

\item{type}{The type of bootstrap procedure. Can be either "nonparametric" for edge testing or
"casedropping" for centrality testing.}

\item{sample_prob}{The vector of sampling proportion for casedropping bootstrap. Default to (0.9,0.8,...0.5).}

\item{nBoots}{The number of bootstraps. Default to 500.}

\item{nCores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.
Defaults to 1.}

\item{verbose}{Logical. Should progress of the function be printed to the console?
Defaults to TRUE.}

\item{library}{Library location to be used in parallel computing. Default to .libPaths().}

\item{seed}{The random seed.}
}
\value{
An IndivBootSplitting object.

\code{IsingResult_new} The new IndivIsing model fitted on the inference dataset.

\code{bootResults} A list containing the bootstraped networks.

\code{id} The input id variable.

\code{type} The input type of bootstrap.

\code{sample_prob} The input sampling proportion for casedropping bootstrap.

\code{time} Time spent on running the function.
}
\description{
This function performs data splitting inference on the constructed individual static or
temporal network using bootstrap to evaluate its accuracy and stability. A nonparametric
bootstrap sampled with replacement on an independent dataset based on the model constructed on the
original dataset is performed to assess the accuracy of estimated network connections.
A case-dropping bootstrap sampled without replacement is performed to
assess the stability of the calculated centrality indices.
}
