% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivBootSummary.R
\name{IndivBootSummary}
\alias{IndivBootSummary}
\title{Summarize the bootstrap results.}
\usage{
IndivBootSummary(
  data,
  IsingBootResult,
  IsingResult,
  nCores = 1,
  library = .libPaths(),
  verbose = TRUE,
  sample_covar_mat
)
}
\arguments{
\item{data}{The data frame containing the binary nodes and the individual characteristics
for network construction.}

\item{IsingBootResult}{The object generated by the IndivBootnet function.}

\item{IsingResult}{The fitted IndivIsing object.}

\item{nCores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.
Defaults to 1.}

\item{library}{Library location to be used in parallel computing. Default to .libPaths().}

\item{verbose}{Logical. Should progress of the function be printed to the console?
Defaults to TRUE.}

\item{sample_covar_mat}{The matrix of covariate information for representative subjects toperform casedropping bootstrap.}
}
\value{
An IndivBootnetSummary object.

\code{boot_data} A matrix containing all the edge weights estimated in each bootstrap replicate.
Only returned for "nonparametric" bootstrap.

\code{boot_data_summary} A matrix containing summary of all bootstrap results. Only returned
for "nonparametric" bootstrap.

\code{edge_data} A matrix containing all edge weights retained in the original constructed network that
estimated in each bootstrap replicate. Only returned for "nonparametric" bootstrap.

\code{edge_data_summary} A matrix containing summary of all edge weights retained in the original constructed network.
Only returned for "nonparametric" bootstrap.

\code{edge_test} A matrix listing the test results of whether the edges are significantly different
from each other. Only the selected edges in the IndivIsing object will be listed.
Only returned for "nonparametric" bootstrap.

\code{edge_quantile} A matrix listing the range of the differences between two estimated edges.
Only the selected edges in the IndivIsing object will be listed. Only returned for "nonparametric" bootstrap.

\code{boot_centrality} A matrix summarizing the centrality information for each subject at each bootstrap.
Only returned for "nonparametric" bootstrap.

\code{test_centrality} A matrix listing the stacked testing results for different centralities between
nodes for the first representative subject.

\code{test_centrality_long} A matrix with the long format listing the stacked testing results for different centralities between
nodes for the first representative subject.

\code{cor_results} Summary of the correlation results. Only returned for "casedropping" bootstrap.

\code{time} Time used for computation.
}
\description{
Summarize the bootstrap results.
}
