% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivIsing.R
\name{IndivIsing}
\alias{IndivIsing}
\title{Build individual-level static or temporal network for binary data}
\usage{
IndivIsing(
  data,
  y_index,
  x_index,
  networkModel = c("static", "temporal"),
  timepoint = NA,
  family = "binomial",
  progressbar = TRUE,
  lowerbound.lambda = NA,
  gamma = 0.25,
  AND = TRUE
)
}
\arguments{
\item{data}{The dataset containing the binary nodes and the individual characteristics
for network construction. If a static network is to be constructed, there is one observation
for each subject; if a temporal network is to be constructed, multiple observations
for each subject is required. Only lag-1 factorization is supported in the current version of
estimating temporal networks.}

\item{y_index}{The indices of the binary nodes.}

\item{x_index}{The indices of the individual characteristics. Pre-processing of the
individual features are required (e.g., scaling of continuous variables, generating
dummy variables for categorical variables).}

\item{networkModel}{Model for network estimation. Can be either 'static' or 'temporal'.}

\item{timepoint}{The column name in 'data' that indicates timepoint for the construction
of temporal network. If \code{networkModel='static'}, \code{timepoint} does not need to be specified. Default to NA.}

\item{family}{Currently only 'binomial' family of data is supportedn for binary nodes.}

\item{progressbar}{Logical. Should the pbar be plotted in order to see the progress
of the estimation procedure? Default to TRUE.}

\item{lowerbound.lambda}{The minimum value of tuning parameter lambda (regularization parameter).
Can be used to compare networks that are based on different sample sizes. The lowerbound.lambda
is based on the number of observations in the smallest group n: sqrt(log(p)/n).
p is the number of variables, that should be the same in both groups. When both
networks are estimated with the same lowerbound for lambda (based on the smallest group),
the two networks can be directly compared. Default to NA.}

\item{gamma}{A value of hyperparameter gamma in the extended BIC. Can be anything
between 0 and 1. Defaults to 0.25.}

\item{AND}{Logical. Can be TRUE of FALSE to indicate whether the AND-rule or
the OR-rule should be used to define the edges in the network. Defaults to TRUE.}
}
\value{
An IndivIsing object.

\code{nodes} The names of nodes in the fitted network.

\code{covar} The names of individual characteristics.

\code{y_index} The input index of the nodes.

\code{x_index} The input index of the individual characteristics.

\code{networkModel} The input model of the network constructed.

\code{timepoint} The input timepoint variable.

\code{AND} The input AND variable.

\code{family} The input family variable.

\code{progressbar} The input progressbar variable.

\code{lowerbound.lambda} The input lowerbound.lambda variable.

\code{gamma} The input gamma variable.

\code{estimated_thresholds} The estimated intercepts in the extended Ising model.

\code{estimated_coeff_raw} The estimated coefficients in the extended Ising model.

\code{estimated_bias} The formula of the estimated bias term, including the estimated
intercept and the coefficients of the major effects of the individual characteristics.

\code{estimated_bias_index} The list of the indices (in the estimated_coeff_raw matrix) of the features
that are involved in the determination of the bias for each node.

\code{estimated_formula} The formula of the estimated edge weights, including the
coefficients of the major effects of the nodes and the coefficients of the interaction terms of the
nodes and the individual characteristics.

\code{estimated_formula_index} The list of the indices (in the estimated_coeff_raw matrix) of the nodes and interaction terms
that are involved in the determination of the edges for each node.

\code{time} Time used for computation.
}
\description{
This function takes the indices of the binary nodes and the indices of the
individual characteristics to estimate an individual-level network based on
the extended Ising or logistic autogressive model. The network can be either
static based on cross-sectional data or temporal based on longitudinal data.
}
