% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivNetwork.R
\name{IndivNetwork}
\alias{IndivNetwork}
\title{Generate individual static or temporal network with binary data for a specific subject.}
\usage{
IndivNetwork(IsingResult, data, id, target_id, covar_vec)
}
\arguments{
\item{IsingResult}{The fitted IndivIsing object.}

\item{data}{The data frame used to fit the IndivIsing object.}

\item{id}{The column name in 'data' that indicates the ID of each subject.}

\item{target_id}{The ID of the specific subject that an individual static or temporal network
is to be fitted.}

\item{covar_vec}{The vector of individual characteristics for a specific subject to plot.
The order has to be the same as x_index. Have to be assigned if target_id is not given.}
}
\value{
An IndivNetwork object.

\code{estimated_network} The estimated adjacent matrix for the specific subject.

\code{net_centrality} A list containing the centrality statistics of the generated graph.

\code{node_centrality} The matrix of the calculated centrality values.

\code{time} Time used for computation.
}
\description{
This function generates the individual static or temporal network with binary nodes
for a specific subject based on the estimated IndivIsing object.
}
