% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivRep.R
\name{IndivRep}
\alias{IndivRep}
\title{Summarize the estimating results based on replicates.}
\usage{
IndivRep(
  data,
  y_index,
  x_index,
  networkModel = c("static", "temporal"),
  timepoint = NA,
  family = "binomial",
  progressbar = TRUE,
  lowerbound.lambda = NA,
  gamma = 0.25,
  AND = TRUE,
  id,
  rep = 200,
  nBoots = 500,
  nCores = 1,
  sample_prob = seq(0.9, 0.5, -0.1),
  verbose = TRUE,
  library = .libPaths(),
  seed,
  sample_covar_mat
)
}
\arguments{
\item{data}{The data frame containing the binary nodes and the individual characteristics
for network construction. If a static network is to be constructed, there is one observation
for each subject. Otherwise, if a temporal network is to be constructed, multiple observations
for each subject is required. Only lag-1 factorization is supported in the current version of
estimating temporal networks.}

\item{y_index}{The indices of the nodes.}

\item{x_index}{The indices of the individual characteristics. Pre-processing of the
individual features are required (e.g., scaling of continuous variables, generating
dummy variables for categorical variables) before performing analysis using this function.}

\item{networkModel}{Model that should be used for network estimation. Can be either 'static' or 'temporal'.}

\item{timepoint}{The column name in 'data' that denotes timepoint for the construction
of temporal network. If \code{networkModel='static'}, \code{timepoint} does not need to be specified.Default to NA.}

\item{family}{Currently only 'binomial' family of data is supported.}

\item{progressbar}{Logical. Should the pbar be plotted in order to see the progress
of the estimation procedure? Default to TRUE.}

\item{lowerbound.lambda}{The minimum value of tuning parameter lambda (regularization parameter).
Can be used to compare networks that are based on different sample sizes. The lowerbound.lambda
is based on the number of observations in the smallest group n: sqrt(log(p)/n).
p is the number of variables, that should be the same in both groups. When both
networks are estimated with the same lowerbound for lambda (based on the smallest group),
the two networks can be directly compared. Default to NA.}

\item{gamma}{A value of hyperparameter gamma in the extended BIC. Can be anything
between 0 and 1. Defaults to 0.25.}

\item{AND}{Logical. Can be TRUE of FALSE to indicate whether the AND-rule or
the OR-rule should be used to define the edges in the network. Defaults to TRUE}

\item{id}{The column name in 'data' that denotes the ID of each subject.}

\item{rep}{The number of replicates of random splitting on the original dataset for network estimation.}

\item{nBoots}{The number of bootstraps. Default to 500.}

\item{nCores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.
Defaults to 1.}

\item{sample_prob}{The vector of sampling proportion for casedropping bootstrap.}

\item{verbose}{Logical. Should progress of the bootstrap function be printed to the console?
Defaults to TRUE.}

\item{library}{Library location to be used in parallel computing.}

\item{seed}{The random seed.}

\item{sample_covar_mat}{The matrix of covariate information for representative subjects to perform casedropping bootstrap.}
}
\value{
An IndivRepSummary object.

\code{list_estimated_thresholds} The list of estimated thresholds based on replicates.

\code{list_estimated_coeff_raw} The list of estimated raw coefficients based on replicates.

\code{list_boot_data} The list of estimated bootstrapped edge values based on replicates.

\code{list_boot_centrality} The list of estimated bootstrapped centrality values based on replicates.

\code{list_cor_final_results} The list of estimated correlations of centralities based on replicates.

\code{data} The input dataset.

\code{y_index} The input indices of nodes.

\code{x_index} The input indices of individual characteristics.

\code{networkModel} The input networkModel parameter.

\code{timepoint} The input timepoint parameter.

\code{family} The input family parameter.

\code{progressbar} The input progressbar parameter.

\code{lowerbound.lambda} The input lowerbound.lambda parameter.

\code{gamma} The input gamma parameter.

\code{AND} The input AND parameter.

\code{rep} The input rep parameter.
}
\description{
Summarize the estimating results based on replicates.
}
