% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivRepSummary.R
\name{IndivRepSummary}
\alias{IndivRepSummary}
\title{Summarize the estimating results based on replicates.}
\usage{
IndivRepSummary(IndivRepResult, cutoff = 0.5)
}
\arguments{
\item{IndivRepResult}{An returned IndivRep object.}

\item{cutoff}{The cutoff of the probability among all the replicatesbeyond which the edge will be
considered to be included in the final estimated network.}
}
\value{
An IndivRepSummary object.

\code{result_rep} An IndivIsing object that contains the network information estimated based on replicates.

\code{boot_data} A matrix containing all the edge weights estimated in each bootstrap replicate.
Only returned for "nonparametric" bootstrap.

\code{boot_data_summary} A matrix containing summary of all bootstrap results. Only returned
for "nonparametric" bootstrap. Only returned for "nonparametric" bootstrap.

\code{edge_data} A matrix containing all edge weights retained in the original constructed network that
estimated in each bootstrap replicate.

\code{edge_data_summary} A matrix containing summary of all edge weights retained in the original constructed network.

\code{edge_test} A matrix listing the test results of whether the edges are significantly different
from each other.

\code{edge_quantile} A matrix listing the range of the differences between two estimated edges.

\code{boot_centrality} A matrix summarizing the centrality information for each subject at each bootstrap.
#' @returns \code{test_centrality} A matrix listing the stacked testing results for different centralities between
nodes for the first representative subject.

\code{test_centrality_long} A matrix with the long format listing the stacked testing results for different centralities between
nodes for the first representative subject.

\code{cor_final_results} A matrix listing the ranges of the correlation of centralities.
}
\description{
Summarize the estimating results based on replicates.
}
